add helper for jwt
